# Parallelize bytes copies over a certain threshold in ZMQ