@bmcdonald3 2021-08-04
After getting some help from Elliot making some changes to my chapcs bashrc, I had my first paratest run where there were no failures and this PR passed the test! My first paratest with no failures (even over the summer) ❤️